requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
django-tinymce==3.6.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tzdata==2023.3
